src/dags/requirements.txt
psycopg
vertica_python
python-dotenv